REFACTOR: pr 피드백 반영
- forEach -> reduce로 구현
- "all", "subscribe" 같이 명확하지 않은 변수명  구체화
- 파일이름, 폴더이름 명사로 변경
