+++ components/subscriptionSwitchBtn/style.css
